fix some bugs

+++ 1/src/gamma.h
@@ -1,6 +1,6 @@
 #include "iRRAM.h"
 
-#define euler_gamma REAL(0.57721566490)
+#define euler_gamma (0.57721566490)
 
 using namespace iRRAM;
